(PDK-1364) Refactor TemplateDir into classes
Previously the TemplateDir was modified to behave differently between a Git
and local filesystem.  This commit refactors TemplateDir into separate classes:
- Base for Common methods
- Git to override for Git specific behaviour
- Local to override for Git specific behaviour

The PDK::Module::TemplateDir is now a module instead of class and uses a factory
style method of new to create the appropriate TemplateDir class

THe tests have also been split into their new respective files.
