+++ spec/unit/pdk/module/template_dir/base_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+require 'yaml'
+require 'pdk/module/template_dir/base'
+
+describe PDK::Module::TemplateDir::Base do
+  it 'has a metadata method' do
+    expect(described_class.instance_methods(false)).to include(:metadata)
+  end
+end
