+++ spec/unit/pdk/module/template_dir_spec.rb
@@ -37,6 +37,10 @@ describe PDK::Module::TemplateDir do
   end
 
   describe '.new' do
+    before(:each) do
+      allow(described_class).to receive(:validate_module_template!).with(uri.shell_path).and_return(true)
+    end
+
     context 'when not passed a block' do
       it 'raises an ArgumentError' do
         expect {
@@ -52,12 +56,6 @@ describe PDK::Module::TemplateDir do
         }.to raise_error(ArgumentError, %r{must be initialized with a PDK::Util::TemplateURI}i)
       end
     end
-  end
-
-  describe '#template_dir_type' do
-    before(:each) do
-      allow(described_class).to receive(:validate_module_template!).with(uri.shell_path).and_return(true)
-    end
 
     context 'with a git based template directory' do
       before(:each) do
@@ -65,8 +63,8 @@ describe PDK::Module::TemplateDir do
         allow(PDK::Util::Git).to receive(:work_tree?).with(uri.shell_path).and_return(true)
       end
 
-      it 'returns TEMPLATE_DIR_GIT_TYPE' do
-        expect(template_dir.template_dir_type).to eq(PDK::Module::TemplateDir::TEMPLATE_DIR_GIT_TYPE)
+      it 'returns a git based template' do
+        expect(template_dir).to be_a(PDK::Module::TemplateDir::Git)
       end
     end
 
@@ -75,8 +73,8 @@ describe PDK::Module::TemplateDir do
         allow(PDK::Util::Git).to receive(:repo?).with(path_or_url).and_return(false)
       end
 
-      it 'returns TEMPLATE_DIR_FILESYSTEM_TYPE' do
-        expect(template_dir.template_dir_type).to eq(PDK::Module::TemplateDir::TEMPLATE_DIR_FILESYSTEM_TYPE)
+      it 'returns a local based template' do
+        expect(template_dir).to be_a(PDK::Module::TemplateDir::Local)
       end
     end
   end
@@ -143,11 +141,13 @@ describe PDK::Module::TemplateDir do
 
       context 'and it specifies an deprecated built-in template' do
         before(:each) do
+          require 'pdk/module/template_dir/git'
+
           # rubocop:disable RSpec/AnyInstance
           allow(PDK::Util).to receive(:package_install?).and_return(true)
           allow(PDK::Util::Filesystem).to receive(:fnmatch?).with(anything, path_or_url).and_return(true)
           allow(PDK::Util).to receive(:package_cachedir).and_return(File.join('/', 'path', 'to', 'package', 'cachedir'))
-          allow_any_instance_of(described_class).to receive(:clone_template_repo).and_return(path_or_url)
+          allow_any_instance_of(PDK::Module::TemplateDir::Git).to receive(:clone_template_repo).and_return(path_or_url)
           allow(PDK::Util::Git).to receive(:repo?).with(path_or_url).and_return(true)
           allow(PDK::Util::Filesystem).to receive(:rm_rf)
           allow(PDK::Util::Git).to receive(:git).with('--git-dir', anything, 'describe', '--all', '--long', '--always', anything).and_return(stdout: 'ref', exit_code: 0)
@@ -169,72 +169,6 @@ describe PDK::Module::TemplateDir do
     end
   end
 
-  describe '#checkout_template_ref' do
-    let(:path) { File.join('/', 'path', 'to', 'workdir') }
-    let(:ref) { '12345678' }
-    let(:full_ref) { '123456789abcdef' }
-
-    before(:each) do
-      # rubocop:disable RSpec/AnyInstance
-      allow_any_instance_of(described_class).to receive(:clone_template_repo).and_return(path)
-      allow(PDK::Util::Git).to receive(:repo?).with(anything).and_return(true)
-      allow(PDK::Util::Filesystem).to receive(:rm_rf).with(path)
-      allow(described_class).to receive(:validate_module_template!)
-      allow(PDK::Util::Git).to receive(:describe).and_return('git-ref')
-      # rubocop:enable RSpec/AnyInstance
-    end
-
-    context 'when the template workdir is clean' do
-      before(:each) do
-        allow(PDK::Util::Git).to receive(:work_dir_clean?).with(path).and_return(true)
-        allow(Dir).to receive(:chdir).with(path).and_yield
-        allow(PDK::Util::Git).to receive(:ls_remote).with(path, ref).and_return(full_ref)
-      end
-
-      context 'and the git reset succeeds' do
-        before(:each) do
-          allow(PDK::Util::Git).to receive(:git).with('reset', '--hard', full_ref).and_return(exit_code: 0)
-        end
-
-        it 'does not raise an error' do
-          expect {
-            template_dir.checkout_template_ref(path, ref)
-          }.not_to raise_error
-        end
-      end
-
-      context 'and the git reset fails' do
-        let(:result) { { exit_code: 1, stderr: 'stderr', stdout: 'stdout' } }
-
-        before(:each) do
-          allow(PDK::Util::Git).to receive(:git).with('reset', '--hard', full_ref).and_return(result)
-        end
-
-        it 'raises a FatalError' do
-          expect(logger).to receive(:error).with(result[:stdout])
-          expect(logger).to receive(:error).with(result[:stderr])
-          expect {
-            template_dir.checkout_template_ref(path, ref)
-          }.to raise_error(PDK::CLI::FatalError, %r{Unable to checkout '12345678' of git repository at '/path/to/workdir'}i)
-        end
-      end
-    end
-
-    context 'when the template workdir is not clean' do
-      before(:each) do
-        allow(PDK::Util::Git).to receive(:work_dir_clean?).with(path).and_return(false)
-      end
-
-      after(:each) do
-        template_dir.checkout_template_ref(path, ref)
-      end
-
-      it 'warns the user' do
-        expect(logger).to receive(:warn).with(a_string_matching(%r{uncommitted changes found}i))
-      end
-    end
-  end
-
   context 'with a valid template path' do
     it 'returns config hash with module metadata' do
       allow(PDK::Util::Filesystem).to receive(:directory?).with(anything).and_return(true)
@@ -248,10 +182,6 @@ describe PDK::Module::TemplateDir do
       allow(described_class).to receive(:new).with(uri, module_metadata).and_yield(template_dir)
       expect(template_dir.object_config).to include('module_metadata' => module_metadata)
     end
-  end
-
-  it 'has a metadata method' do
-    expect(described_class.instance_methods(false)).to include(:metadata)
   end
 
   describe '.files_in_template(dirs)' do
